requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.2
cffi==1.17.1
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.115.12
fastapi-crudrouter==0.8.6
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
pycparser==2.22
pydantic==1.10.22
pydantic_core==2.33.2
PyJWT==2.10.1
PyMySQL==1.1.1
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.2
websockets==15.0.1
cryptography